requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cloudinary==1.33.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
defusedxml==0.7.1
Deprecated==1.2.14
dj-rest-auth==4.0.1
Django==4.2.3
django-allauth==0.54.0
django-appconf==1.0.5
django-cloudinary-storage==0.3.0
django-compressor==4.4
django-cors-headers==4.2.0
django-filter==23.2
django-oauth-toolkit==2.3.0
django-rest-passwordreset==1.3.0
django-templated-mail==1.1.1
django-webpack-loader==2.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-social-oauth2==2.1.3
ecdsa==0.18.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.0
Markdown==3.4.3
MarkupSafe==2.1.3
oauthlib==3.2.2
Pillow==10.0.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.7.0
pylibmagic==0.4.0
python-decouple==3.8
python-jose==3.3.0
python-magic==0.4.27
python3-openid==3.2.0
pytz==2023.3
rcssmin==1.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.16
whitenoise==6.5.0
wrapt==1.15.0
